requirements.txt
discord.py
requests
discord-py-interactions
riotwatcher
openai>=1.0.0
psycopg2-binary>=2.9.0